moves logging functions to new file log_utils.py
